refactor: add useEventStreamListener hook
