+++ src/modules/event-stream/EventStreamContext.tsx
@@ -1,22 +1,21 @@
-import {createContext, useContext, useEffect} from 'react';
+import {createContext, useContext} from 'react';
 import {
   StreamEventListener,
   useSetupEventStream,
 } from './use-setup-event-stream';
 import {EventKind, StreamEventLog} from './types';
-import {useQueryClient} from '@tanstack/react-query';
-import {getBonusBalanceQueryKey} from '../bonus/queries/use-bonus-balance-query';
-import {getBonusAmountEarnedQueryKey} from '../bonus';
-import {useAuthContext} from '../auth';
-import {getShmoBookingQueryKey} from '../mobility/queries/use-shmo-booking-query';
-import {languageGlobal} from '../locale';
-import {getActiveShmoBookingQueryKey} from '../mobility/queries/use-active-shmo-booking-query';
-import {invalidateFareContractsQuery} from '../ticketing/use-fare-contracts';
-import {useFeatureTogglesContext} from '../feature-toggles';
 
 interface EventStreamContextValue {
   eventLog: StreamEventLog;
+  /**
+   * NOTE: Usually should not be called directly, but rather through the
+   * `useEventStreamListener` hook.
+   */
   subscribe: <K extends EventKind>(listener: StreamEventListener<K>) => void;
+  /**
+   * NOTE: Usually should not be called directly, but rather through the
+   * `useEventStreamListener` hook.
+   */
   unsubscribe: (id: string) => void;
 }
 
@@ -42,73 +41,6 @@ type Props = {
 
 export const EventStreamContextProvider = ({children}: Props) => {
   const {eventLog, subscribe, unsubscribe} = useSetupEventStream();
-  const queryClient = useQueryClient();
-  const {userId} = useAuthContext();
-  const {isEventStreamFareContractsEnabled} = useFeatureTogglesContext();
-
-  useEffect(() => {
-    const fareContractListenerId = 'fare-contract-listener';
-    subscribe({
-      id: fareContractListenerId,
-      eventKind: EventKind.FareContract,
-      callback: () => {
-        if (isEventStreamFareContractsEnabled) {
-          invalidateFareContractsQuery(queryClient);
-        }
-      },
-    });
-    return () => {
-      unsubscribe(fareContractListenerId);
-    };
-  }, [subscribe, unsubscribe, queryClient, isEventStreamFareContractsEnabled]);
-
-  useEffect(() => {
-    const shmoBookingListenerId = 'shmo-booking-updated-listener';
-    subscribe({
-      id: shmoBookingListenerId,
-      eventKind: EventKind.ShmoBookingUpdated,
-      callback: (streamEvent) => {
-        queryClient.invalidateQueries({
-          queryKey: getActiveShmoBookingQueryKey(languageGlobal),
-        });
-        queryClient.invalidateQueries({
-          queryKey: getShmoBookingQueryKey(
-            streamEvent.bookingId,
-            languageGlobal,
-          ),
-        });
-      },
-    });
-    return () => {
-      unsubscribe(shmoBookingListenerId);
-    };
-  }, [
-    subscribe,
-    unsubscribe,
-    queryClient,
-    userId,
-    isEventStreamFareContractsEnabled,
-  ]);
-
-  useEffect(() => {
-    const personalisationProgramPointListenerId =
-      'personalisation-program-point-listener';
-    subscribe({
-      id: personalisationProgramPointListenerId,
-      eventKind: EventKind.PersonalisationProgramPoint,
-      callback: (streamEvent) => {
-        queryClient.invalidateQueries({
-          queryKey: getBonusBalanceQueryKey(userId),
-        });
-        queryClient.invalidateQueries({
-          queryKey: getBonusAmountEarnedQueryKey(userId, streamEvent.orderId),
-        });
-      },
-    });
-    return () => {
-      unsubscribe(personalisationProgramPointListenerId);
-    };
-  }, [subscribe, unsubscribe, queryClient, userId]);
 
   return (
     <EventStreamContext.Provider
