+++ src/modules/event-stream/index.ts
@@ -1,2 +1,5 @@
 export {useSetupEventStream} from './use-setup-event-stream';
 export {EventStreamContext, useEventStreamContext} from './EventStreamContext';
+export {EventKind} from './types';
+export {useEventStreamListener} from './use-event-stream-listener';
+export {useGlobalEventStreamListeners} from './use-global-event-stream-listeners';
